refactor(build): replace pixi build with deprecation shim, move logic to pixi publish
`pixi build` is replaced with a thin deprecation shim that prints a
warning and forwards to `pixi publish` with `target_dir = output_dir`.
The full build pipeline (`validate_package_manifest`,
`determine_discovery_start`, and the package build / artifact
collection logic) moves to `pixi publish`, which now handles
build + upload in a single command.

- Strip the build pipeline out of `crates/pixi_cli/src/build.rs`,
  leaving the `Args` struct intact and replacing `execute` with a
  shim that prints a `pixi build is deprecated` warning, renders the
  equivalent `pixi publish` invocation, and delegates to
  `publish::execute`.
- Update the `pixi build` entry in the frozen/no-install invariant
  loop to expect failure (the publish entry already exercises the
  happy path).
- Convert `test_workspace_variants_separate_work_directories` to use
  `pixi publish --to file://` so variant work-directory separation
  remains tested.
- Rename `crates/pixi/tests/integration_rust/build_tests.rs` to
  `source_package_tests.rs` and drop three tests whose coverage no
  longer applies after the move:
  - `test_build_command_with_recipe_yaml_path`: the assertions only
    exercised `fs::write` / `fs::read_to_string` and never invoked
    pixi.
  - `test_build_creates_gitignore`: regression test for #4761; the
    `.pixi/.gitignore` invariant is still upheld because pixi
    publish calls the same `sanity_check_workspace`. A publish-side
    replacement is added in a follow-up commit.
  - `test_build_fails_before_any_build_when_one_variant_is_unsatisfiable`:
    the publish equivalent
    `test_publish_fails_before_build_or_upload_when_one_variant_is_unsatisfiable`
    already covers this.
