+++ crates/pixi_cli/src/task.rs
@@ -300,7 +300,7 @@ fn print_tasks(
         header_style.apply_to("Task"),
         header_style.apply_to("Description"),
     );
-    writeln!(writer, "{}", &header)?;
+    writeln!(writer, "{}", header)?;
     for (taskname, row) in formatted_descriptions {
         writeln!(writer, "{}\t{}", taskname.fancy_display(), row)?;
     }
