+++ tests/integration_python/pixi_build/test_ros.py
@@ -36,14 +36,7 @@ def test_ros_packages_build(
     manifest_path = workspace.joinpath("src", package_dir, "pixi.toml")
 
     verify_cli_command(
-        [
-            pixi,
-            "build",
-            "--path",
-            manifest_path,
-            "--output-dir",
-            output_dir,
-        ]
+        [pixi, "publish", str(output_dir), "--path", manifest_path],
     )
 
     expected_name = ROS_PACKAGE_OUTPUT_NAMES[package_dir]
@@ -65,14 +58,7 @@ def test_ros_packages_build_point_to_package_xml(
     manifest_path = workspace.joinpath("src", package_dir, "package.xml")
 
     verify_cli_command(
-        [
-            pixi,
-            "build",
-            "--path",
-            manifest_path,
-            "--output-dir",
-            output_dir,
-        ]
+        [pixi, "publish", str(output_dir), "--path", manifest_path],
     )
 
     expected_name = ROS_PACKAGE_OUTPUT_NAMES[package_dir]
@@ -93,14 +79,7 @@ def test_ros_packages_build_point_to_package_xml_in_the_same_dir(
     manifest_path = workspace.joinpath("src", ROS_IMPLICIT_PACKAGE_DIR)
 
     verify_cli_command(
-        [
-            pixi,
-            "build",
-            "--path",
-            "package.xml",
-            "--output-dir",
-            output_dir,
-        ],
+        [pixi, "publish", str(output_dir), "--path", "package.xml"],
         cwd=manifest_path,
     )
 
@@ -120,14 +99,7 @@ def test_ros_packages_build_point_to_implicit_package_xml_fails(
     manifest_path = workspace.joinpath("src", ROS_IMPLICIT_PACKAGE_DIR)
 
     verify_cli_command(
-        [
-            pixi,
-            "build",
-            "--path",
-            manifest_path,
-            "--output-dir",
-            output_dir,
-        ],
+        [pixi, "publish", str(output_dir), "--path", manifest_path],
         expected_exit_code=ExitCode.FAILURE,
         stderr_contains=[
             "is a directory, please provide the path to the manifest file",
