+++ tests/integration_python/pixi_build/test_variants.py
@@ -31,14 +31,7 @@ def test_inline_variants_produce_multiple_outputs(
     output_dir = tmp_pixi_workspace.joinpath("dist")
 
     verify_cli_command(
-        [
-            pixi,
-            "build",
-            "--path",
-            manifest_path,
-            "--output-dir",
-            output_dir,
-        ],
+        [pixi, "publish", str(output_dir), "--path", manifest_path],
     )
 
     # Ensure that we don't create directories we don't need
@@ -125,14 +118,7 @@ def test_variant_files_produce_multiple_outputs(
     output_dir = tmp_pixi_workspace.joinpath("dist")
 
     verify_cli_command(
-        [
-            pixi,
-            "build",
-            "--path",
-            manifest_path,
-            "--output-dir",
-            output_dir,
-        ],
+        [pixi, "publish", str(output_dir), "--path", manifest_path],
     )
 
     built_packages = list(output_dir.glob("*.conda"))
